requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
django-filter==23.1
dotenv
Pillow==9.4.0
PyTelegramBotAPI
requests==2.28.2